refactor: 사용하지 않는 클래스 제거
