--- src/main/java/com/gugucon/shopping/pay/dto/toss/response/TossPayCreateResponse.java
@@ -1,19 +1,0 @@
-package com.gugucon.shopping.pay.dto.toss.response;
-
-import com.gugucon.shopping.pay.domain.Pay;
-import lombok.AccessLevel;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@AllArgsConstructor(access = AccessLevel.PRIVATE)
-@NoArgsConstructor(access = AccessLevel.PRIVATE)
-@Getter
-public class TossPayCreateResponse {
-
-    private Long payId;
-
-    public static TossPayCreateResponse from(final Pay pay) {
-        return new TossPayCreateResponse(pay.getId());
-    }
-}
